Fixed BattleCharacter's getHealth and mana methods returning floats Added a status() method to Party that returns a string describing the HP and mana of each party member, separated by \n Shortened the Battle set actions dialogue Moved the conditional tests for fainted sources and destinations to BattleAction.run() Updated BattleRenderer to render main party status CLEAR_DIALOGUE now also overrides attach: it checks if the attached is also a CLEAR_DIALOGUE, in which case it just returns it, otherwise it returns super.attach(builder) Changed the game height to 352 so it is an exact multiple of 32 The font is now scaled to be 3/4 of a grid space (24 px). This way, including frame games, 12 lines fit exactly on the screen DialogueStyles have been updated accordingly Fixed the mysterious lack of transparency with dialogue windows Improved EffectParser and Effect Implemented Effect in Item class Removed the MENU state from GameScreen, and removed setBattle, putting it's funcitonality in battle() (formerly goToBattle)

+++ src/com/arnopaja/supermac/battle/characters/BattleCharacter.java
@@ -72,8 +72,8 @@ public abstract class BattleCharacter {
     public String getName() { return name; }
     public BattleClass getBattleClass() { return battleClass; }
     public int getLevel() { return level; }
-    public float getMaxHealth() { return maxHealth; }
-    public float getMaxMana() { return maxMana; }
+    public int getMaxHealth() { return maxHealth; }
+    public int getMaxMana() { return maxMana; }
     public int getAttack() { return attack; }
     public int getDefense() {
         if(isDefending) return defense * 2;
@@ -81,8 +81,8 @@ public abstract class BattleCharacter {
     }
     public int getSpecial() { return special; }
     public int getSpeed() { return speed; }
-    public float getHealth() { return currentHealth; }
-    public float getMana() { return currentMana; }
+    public int getHealth() { return currentHealth; }
+    public int getMana() { return currentMana; }
     public Armor getEquippedArmor() { return equippedArmor; }
     public Weapon getEquippedWeapon() { return equippedWeapon; }
     public List<Spell> getSpellsList() { return spellsList; }
